--- MystatDesktopWpf/UserControls/Debug.xaml
@@ -1,34 +1,0 @@
-﻿<UserControl x:Class="MystatDesktopWpf.UserControls.Debug"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:MystatDesktopWpf"
-             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
-             xmlns:domain="clr-namespace:MystatDesktopWpf.Domain"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800"
-             Background="{DynamicResource MaterialDesignPaper}">
-    <WrapPanel>
-        <materialDesign:Card Margin="16" Padding="24">
-            <StackPanel>
-                <Button Content="Snackbar" Margin="8" Click="Button_Snackbar_Click" HorizontalAlignment="Left"/>
-                <Button Content="Snackbar with 10 seconds timer" Margin="8" Click="Button_SnackbarTimer_Click" HorizontalAlignment="Left"/>
-                <Button Content="Noification dialog " Margin="8" Click="Button_Notification_Click" HorizontalAlignment="Left"/>
-                <Button Content="Noification dialog delayed" Margin="8" Click="Button_NotificationDelayed_Click" HorizontalAlignment="Left"/>
-            </StackPanel>
-        </materialDesign:Card>
-        <materialDesign:Card Margin="0,16,16,16" Padding="24">
-            <StackPanel>
-                <Label Content="{DynamicResource m_Login}" Margin="8"/>
-                <Button Content="Change to EN" Margin="8" Click="Button_EN_Click"/>
-                <Button Content="Change to RU" Margin="8" Click="Button_RU_Click"/>
-            </StackPanel>
-            
-        </materialDesign:Card>
-        
-        <materialDesign:Snackbar x:Name="notifySnackbar" MessageQueue="{materialDesign:MessageQueue}"
-                                 VerticalAlignment="Bottom" HorizontalAlignment="Right" Margin="32"
-                                 VerticalContentAlignment="Bottom" Grid.RowSpan="3"/>
-    </WrapPanel>
-</UserControl>
