--- MystatDesktopWpf/UserControls/MainMenu.xaml
@@ -1,72 +1,0 @@
-﻿<UserControl x:Class="MystatDesktopWpf.UserControls.MainMenu"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:MystatDesktopWpf"
-             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
-             xmlns:domain="clr-namespace:MystatDesktopWpf.Domain"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800"
-             Background="{DynamicResource MaterialDesignPaper}">
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition/>
-        </Grid.RowDefinitions>
-        <Grid x:Name="TopBar" Background="{DynamicResource MaterialDesignCardBackground}">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="{Binding ElementName=NagivationBar, Path=ActualWidth}"/>
-                <ColumnDefinition/>
-                <ColumnDefinition Width="Auto"/>
-            </Grid.ColumnDefinitions>
-            <Grid.RowDefinitions>
-                <RowDefinition/>
-                <RowDefinition Height="Auto"/>
-            </Grid.RowDefinitions>
-            <ToggleButton Margin="0,5,0,5" Style="{DynamicResource MaterialDesignHamburgerToggleButton}" HorizontalAlignment="Center"/>
-            <Button Grid.Column="2" Margin="0,0,16,0" Click="Button_Exit_Click"
-                      materialDesign:RippleAssist.Feedback="{Binding RelativeSource={RelativeSource Self}, Path=Foreground, Converter={StaticResource BrushRoundConverter}}"
-                      Content="{materialDesign:PackIcon Kind=ExitToApp, Size=24}"
-                      Foreground="{Binding RelativeSource={RelativeSource AncestorType={x:Type FrameworkElement}}, Path=(TextElement.Foreground)}"
-                      Style="{StaticResource MaterialDesignToolButton}"
-                      ToolTip="{DynamicResource m_ExitPopUp}" />
-            <Separator Grid.Column="1" Grid.Row="1" Grid.ColumnSpan="2" Visibility="Collapsed"/>
-            <ComboBox 
-                  Grid.Column="1"
-                  Style="{StaticResource MaterialDesignFilledComboBox}" Margin="0,0,16,0" HorizontalAlignment="Right" Width="auto" x:Name="langsComboBox" >
-            </ComboBox>
-        </Grid>
-
-        <Grid Grid.Row="2">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="Auto"/>
-                <ColumnDefinition/>
-            </Grid.ColumnDefinitions>
-            <Border Background="{DynamicResource MaterialDesignCardBackground}"/>
-            <ListBox VerticalAlignment="Top" x:Name="NagivationBar"
-                     ItemsSource="{Binding MenuItems}"
-                     SelectedValue="{Binding SelectedItem}"
-                     Margin="0,8,0,0">
-                <ListBox.ItemContainerStyle>
-                    <Style TargetType="ListBoxItem" BasedOn="{StaticResource MaterialDesign3.NavigationRailListBoxItem}">
-                        <Setter Property="Background" Value="{DynamicResource PrimaryHueLightBrush}"/>
-                        <Setter Property="materialDesign:BadgedAssist.Badge" Value="{Binding Notifications}" />
-                        <Setter Property="materialDesign:BadgedAssist.BadgeBackground" Value="#930006" />
-                        <Setter Property="materialDesign:BadgedAssist.BadgeForeground" Value="#FFDAD4" />
-                        <Setter Property="materialDesign:NavigationRailAssist.IsTextVisible" Value="True" />
-                        <Setter Property="materialDesign:NavigationRailAssist.SelectedIcon" Value="{Binding SelectedIcon}" />
-                        <Setter Property="materialDesign:NavigationRailAssist.UnselectedIcon" Value="{Binding UnselectedIcon}" />
-                    </Style>
-                </ListBox.ItemContainerStyle>
-                <ListBox.Resources>
-                    <DataTemplate DataType="{x:Type domain:MainMenuItem}">
-                        <TextBlock Text="{Binding Name}" TextAlignment="Center" TextWrapping="Wrap" />
-                    </DataTemplate>
-                </ListBox.Resources>
-            </ListBox>
-            <UserControl DataContext="{Binding SelectedItem}" Grid.Column="1" Content="{Binding Content}"/>
-        </Grid>
-    </Grid>
-</UserControl>
