--- MystatDesktopWpf/UserControls/MainMenu.xaml.cs
@@ -1,61 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-using MaterialDesignThemes.Wpf;
-using MaterialDesignThemes.Wpf.Transitions;
-using MystatDesktopWpf.Domain;
-using MystatDesktopWpf.Services;
-
-namespace MystatDesktopWpf.UserControls
-{
-    /// <summary>
-    /// Interaction logic for MainMenuUserControl.xaml
-    /// </summary>
-    public partial class MainMenu : UserControl
-    {
-        public MainMenu()
-        {
-            this.DataContext = new MainMenuViewModel();
-            InitializeComponent();
-
-            int index = 0;
-            foreach (var lang in App.Languages)
-            {
-                var langItem = new ComboBoxItem();
-                langItem.Content = lang.TwoLetterISOLanguageName.ToUpper();
-                langItem.Tag = (index++).ToString();
-                langItem.IsSelected = lang.Name == App.Language.Name;
-                langItem.Selected += OnLanguageChange;
-                langsComboBox.Items.Add(langItem);
-            }
-        }
-
-        private void Button_Exit_Click(object sender, RoutedEventArgs e)
-        {
-            // Back to login
-            SettingsService.RemoveUserData();
-            Transitioner.MovePreviousCommand.Execute(null, null);
-        }
-
-        private void OnLanguageChange(object sender, RoutedEventArgs e)
-        {
-            var item = sender as ComboBoxItem;
-
-            if (item is null) return;
-
-            var index = int.Parse(item.Tag as string);
-            App.Language = App.Languages[index];
-        }
-    }
-}
