--- MystatDesktopWpf/UserControls/Placeholder.xaml.cs
@@ -1,35 +1,0 @@
-﻿using MaterialDesignThemes.Wpf;
-using MaterialDesignThemes.Wpf.Transitions;
-using MystatAPI.Entity;
-using MystatDesktopWpf.Domain;
-using System;
-using System.Collections.Generic;
-using System.Data.Common;
-using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-
-namespace MystatDesktopWpf.UserControls
-{
-    /// <summary>
-    /// Interaction logic for PlaceholderUserControl.xaml
-    /// </summary>
-    public partial class Placeholder : UserControl
-    {
-        public Placeholder()
-        {
-            InitializeComponent();
-        }
-    }
-}
